% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Returns information of a uri}
\usage{
label(base, uri)
}
\arguments{
\item{base}{The base url of the rst endpoint}

\item{The}{uri of a term}
}
\value{
json representation as list of the information
}
\description{
Returns information of a uri
}

